requirements.txt
absl-py==0.7.1
argcomplete==1.9.3
astor==0.8.0
boto3==1.9.176
botocore==1.12.176
certifi==2019.6.16
cfn-flip==1.2.1
chardet==3.0.4
Click==7.0
colorama==0.3.9
docutils==0.14
durationpy==0.5
Flask==1.0.3
future==0.16.0
gast==0.2.2
gevent==1.4.0
google-pasta==0.1.7
greenlet==0.4.15
grpcio==1.21.1
h5py==2.9.0
hjson==3.0.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.3
kappa==0.6.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
lambda-packages==0.20.0
Markdown==3.1.1
MarkupSafe==1.1.1
numpy==1.16.4
placebo==0.9.0
protobuf==3.8.0
pyasn1==0.4.5
python-dateutil==2.6.1
python-slugify==1.2.4
PyYAML==5.1
requests==2.22.0
rsa==3.4.2
s3transfer==0.2.1
scipy==1.3.0
six==1.12.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
toml==0.10.0
tqdm==4.19.1
troposphere==2.4.8
Unidecode==1.1.1
urllib3==1.25.3
Werkzeug==0.15.4
wrapt==1.11.2
wsgi-request-logger==0.4.6
zappa==0.48.2
